### 2021-08-22 #3644 @whedon
[PRE REVIEW]: QXTools: A Julia framework for distributed quantum circuit simulation
**Submitting author:** @nmoran (<a href="http://orcid.org/0000-0002-2619-5040">Niall Moran</a>)
**Repository:** <a href="https://github.com/JuliaQX/QXTools.jl" target ="_blank">https://github.com/JuliaQX/QXTools.jl</a>
**Version:** 0.2.3
