### 2021-08-22 #3644 @danielskatz
@nmoran - where is the JOSS paper in your repository?
See https://joss.readthedocs.io/en/latest/submitting.html#typical-paper-submission-flow and https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
