### 2021-08-22 #3644 @nmoran
@danielskatz apologies I see that I gave the URL to the repository with the software and not the paper. Paper is at [https://github.com/nmoran/juliaqx_joss](https://github.com/nmoran/juliaqx_joss).

Is it possible to change the URL or do I need to re-submit?
