### 2021-08-22 #3644 @danielskatz
Ideally you would put the paper in the software repository, at least in a branch
