+++ test/runtests.jl
@@ -1,7 +1,0 @@
-using QXTools
-using Test
-using TestSetExtensions
-
-@testset ExtendedTestSet "All QXTools.jl tests" begin
-    @includetests ARGS
-end
