### 2021-08-22 3da8be8
Remove redundant CI configs
