+++ .gitlab/merge_request_templates/default.md
@@ -1,24 +1,0 @@
-### Summary
-
-Short summary of the changes.
-
-### Rationale
-
-Explain the rationale for the changes (links to relevant issue(s)).
-
-#### Implementation Details
-
-#### Additional notes
-
-<hr>
-
-***Check before merging:***
-
-- [ ] All discussions are resolved
-- [ ] New code is covered by appropriate tests
-- [ ] Tests are passing locally and on CI
-- [ ] The documentation is consistent with changes
-- [ ] Any code that was copied from other sources has the paper/url in a comment and is compatible with the MIT licence
-- [ ] Notebooks/examples not covered by unittests have been tested and updated as required
-- [ ] The feature branch is up-to-date with the master branch (rebase if behind)
-- [ ] Incremented the version string in Project.toml file
