+++ .gitlab-ci.yml
@@ -1,88 +1,0 @@
-stages:
-  - test
-  - docs
-
-.script:
-  before_script:
-    ##
-    ## Install ssh-agent if not already installed, it is required by Docker.
-    ## (change apt-get to yum if you use an RPM-based image)
-    ##
-    - 'command -v ssh-agent >/dev/null || ( apt-get update -y && apt-get install openssh-client -y )'
-
-    ##
-    ## Run ssh-agent (inside the build environment)
-    ##
-    - eval $(ssh-agent -s)
-
-    ##
-    ## Add the SSH key stored in SSH_PRIVATE_KEY variable to the agent store
-    ## We're using tr to fix line endings which makes ed25519 keys work
-    ## without extra base64 encoding.
-    ## https://gitlab.com/gitlab-examples/ssh-private-key/issues/1#note_48526556
-    ##
-    - echo "$GITLAB_SSH_PRIVATE_KEY" | tr -d '\r' | ssh-add -
-
-    ##
-    ## Create the SSH directory and give it the right permissions
-    ##
-    - mkdir -p ~/.ssh
-    - chmod 700 ~/.ssh
-    - ssh-keyscan git.ichec.ie >> ~/.ssh/known_hosts
-  script:
-    - |
-      julia --project=@. -e '
-        using Pkg
-        Pkg.build()
-        Pkg.test(coverage=true)'
-.coverage:
-  coverage: /Test coverage (\d+\.\d+%)/
-  after_script:
-    - |
-      julia -e '
-        using Pkg
-        Pkg.add("Coverage")
-        using Coverage
-        c, t = get_summary(process_folder())
-        using Printf
-        @printf "Test coverage %.2f%%\n" 100c / t'
-.docs:
-  script:    
-    - |
-      julia --project -e '
-        using Pkg
-        Pkg.instantiate()
-        Pkg.add("Documenter")
-        Pkg.add("DataFrames")
-        Pkg.add("Latexify")'
-    - julia --project docs/make.jl
-
-Julia 1.5:
-  stage: test
-  image: julia:1.5
-  tags:
-    - docker
-  extends:
-    - .script
-    - .coverage
-
-Julia 1.6:
-  stage: test
-  image: julia:1.6
-  tags:
-    - docker
-  extends:
-    - .script
-    - .coverage
-
-docs:
-  stage: docs
-  image: julia:1.5
-  tags:
-    - docker
-  extends:
-    - .script
-    - .docs
-  artifacts:
-    paths:
-      - docs/build
