### 2021-09-09 #3711 @jarvist
Thanks for agreeing to review Michael and Utkarsh! I hope working to a deadline of two weeks will be feasible?

The three 'reviewer' sections on the online documentation are probably the single best place to refer to while doing the review.
https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

If you come up with any questions or difficulties with installation or understanding what the package does, this is very useful information to feed back to the developers. Small issues are probably best discussed here, but if you have a long list of changes suggested you may want to open up dedicated issues on the software repository.
