### 2021-09-09 #3711 @goerz
Yeah, I can definitely give feedback within the next two weeks.

Just a small meta-comment (is there an issue tracker for that?): For the [review checklist](https://github.com/openjournals/joss-reviews/issues/3711#issue-992554681), it would be convenient for the review lists of the different reviewers to be in a different comment. Last paper I reviewed, I accidentally checked off items for the other reviewer. If they're separate comments, I can bookmark a link to just my checklist.
