### 2021-09-09 #3711 @jarvist
OK, Utkarsh is busy this week, so let's work to a Sept 30th deadline. (Of course, any time before is also great!)

Thank you for the meta comment. I have also managed to do this while reviewing. I'll take it to the JOSS Editor's Slack - I think there's a channel for Whedon development.
