### 2021-09-10 #3711 @jarvist
> Just a small meta-comment...

After brief discussion on the Slack, the answer is that yes there is an issue tracker for such things, and I've added an issue & will attempt to keep track of it:
https://github.com/openjournals/joss/issues/988
