### 2021-09-25 JuliaQX/QXTools.jl#31 @obliviateandsurrender
[JOSS Review] @obliviateandsurrender Some Issues and Suggestions
Dear authors,

Thank you for your work. Most of it looks good to me, but it would be great if you could address the following suggestions and points as part of the  [JOSS submission Review Process](https://github.com/openjournals/joss-reviews/issues/3711).

1. JOSS Paper
    - [x]  Add a reference for `tensor networks` and explain how it works.
    - [x] Mention how it is better or different from other packages (if any) that supposedly intend to quantum circuit simulations in a similar fashion either in Julia or other languages.
    - [x] Mention if there are any limitations of this package.

2. Repository
    - [x]  There definitely needs to be some tutorials (or documentation) for the following things:
    i. Explaining details of the workflow step by step as already mentioned in #5.
    ii. Provide the performance benchmarks for the package (see the point below).
    - [x]  While looking at the repository and the documentation, I could not see any particular examples or relevant results for the following statement.
      > To reduce the turnaround time and allow larger systems to be simulated, it is necessary to move beyond single workstations and use distributed compute clusters.

      While it being obvious, it would be beneficial if some useful examples  (or even better if some tutorials) could be added to showcase how this package reduces simulation time.
    - [x]  Add the link of the documentation explicitly in the `README`. It might not be evident to the user to access it via badge.
    - [x]  [**Suggestion**] I had a look at the contributing guidelines and the pull request template, and they look good! However, I would suggest you have a template for Issues as well, making the user's task a bit easier to request new features or report any bugs.

Remember that all issues are open to discussion, and additional considerations are welcome!
