### 2021-09-27 #3711 @goerz
I've had a look at the software, and verified that it can be installed and works as described. The source code appears to be of high quality, implementing non-trivial methods for quantum circuit simulation, which is an important field of research. Thus, the software package itself appears to be at a state of development that warrants publication. However, the documentation and the submitted JOSS paper are completely insufficient at this time. **Thus, I cannot support publication**. However, I believe that if the authors substantially improve the documentation, a revised paper would be publishable in JOSS. Presumably, this would not require any changes to the software itself.

## Software paper

The description of the purpose, scope, and limitations of the package is insufficient.

My recommendation would be to assume as a target audience a student or researcher who has basic familiarity with quantum circuits (as in, they've read a bit of Nielsen & Chuang or have gone through some Qiskit tutorials). I would not assume knowledge of tensor network methods or matrix product states. The JOSS paper, and the first page of the documentation, should give me an "elevator pitch", allowing me to evaluate whether to consider using QXTools for some project.

With this in mind, first, explain the *purpose* of the package. "Simulating quantum circuits" is too vague. What does that mean, exactly? Just running a Hilbert space state through an ideal circuit once? Classically noisy gates? Decoherence? Parametric quantum circuits? Statistical sampling of measurements?

Second, what are the distinguishing features of QXTools? Why would I want to use it?  It seems like there are two distinguishing features: the use of tensor network methods, and performance on distributed HPC clusters and/or GPU environments. Please give at least a one-paragraph outline of how tensor network methods work and how they affect the simulation of quantum circuits. This probably ties in with the purpose of the package: For a circuit with less than 10 qubits, I can probably do "circuit simulation" just with basic linear algebra directly in Julia, and probably wouldn't even look for a specialized package (except maybe for some kind of extremely user-friendly API). For what kind of circuit and circuit size is QXTools primarily intended? In the JOSS paper, the "Statement of Need" should reflect this information, as well as the "State of the field", which is mostly missing: How does QXTools compare to other packages? The Readme mentions

> The design and implementation of QXTools and related packages was inspired by many other frameworks and packages including ITensors.jl, TensorOperations.jl, Yao.jl, TAL-SH and ExaTN.

You will have to discuss how QXTools compares to these packages. There are also packages outside of the Julia ecosystem, e.g. Qiskit, that would be relevant.

Lastly, please discuss the limitations of QXTools: To how big a circuit can I push the package before simulation becomes infeasible? What are the memory requirements that I might run into? How does this compare to other packages?

## Documentation

* There is a `CONTRIBUTING.md` file, but the link from the README is broken
* "Statement of Needs" is missing from the documentation
* The Example in the Readme is insufficient. At least in the full documentation, you will have to expand this. What is the circuit diagram for a GHZ circuit? What is the corresponding tensor network circuit, and how does it get calculated? What's a "contraction plan", conceptually? What are we actually simulating? This example (3 qubits) seems trivially small, which might be a good thing as it might allow explaining the tensor network circuit etc., going through the calculation "by hand". It might also be good to have an example that's more realistic for the intended usage of QXTools. In a poster video from one of Julia conferences you seem to be running a simulation for a 60 qubit circuit on a laptop, which seems *very* non-trivial to me. So that might a good example to illustrate how the tensor network approach allows you to handle large circuits. Running on an HPC cluster is probably hard to describe in any reproducible way, but that's something where a screencast video could be useful, as an illustration of a possible workflow (totally optional).
* There is a section "Tutorials" in the documentation that is completely blank. I would strongly recommend adding several tutorials that illustrate the usage of the package.
* I would recommend having a single coherent documentation for the QXTools package that includes the documentation for the various sub-packages in the JuliaQX organization (EXZoo, QXContexts, etc.), including explanations of how everything fits together. The sub-packages might then only have a basic skeleton documentation that includes the auto-generated API but refers to the main documentation for explanations / tutorials / etc.

## Functionality

* Performance: There are no performance benchmarks in the documentation. While not strictly necessary for publication in JOSS, performance on distributed HPC clusters is a primary advertised feature of the package, which would warrant some benchmarking analysis, including an analysis of parallelization overhead.

## Other issues

* A small matter of style: I don't think "Julialang" is generally used as the name of the Julia language (except for as a Twitter hashtag)
