### 2021-09-27 #3711 @goerz
@jarvist Done. If you (and the authors) think that the documentation can be brought up to speed within a reasonable timeframe, I'd be happy to review a revised version.
