### 2021-09-28 #3711 @obliviateandsurrender
@jarvist I had left my suggestions for the authors to address in the issue [here](https://github.com/JuliaQX/QXTools.jl/issues/31), a few days ago. These are majorly based on the issues regarding the absence of performance data for the package, the lack of extensive documentation, and introductory tutorials, as pointed out in even greater detail by @goerz. I believe this package could be considered for publication in JOSS only **after** authors make significant improvements on the issues mentioned above.
