### 2021-10-04 #3711 @jarvist
@nmoran , have you had time to consider a response?
I appreciate that writing documentation is a large and quite open ended task, but I think the reviewer's concerns are bona fide, and that better documentation would improve the package in general.
