### 2021-10-05 #3711 @nmoran
First I'd like to thank @obliviateandsurrender and @goerz for taking the time to review our work and provide such detailed feedback. We have been planning to write some tutorial material, but have not managed to get to it yet. Further performance testing is underway also so we expect to have some updated results shortly which we can include. We expect to be able to have both these done within the next two months if this would be reasonable?
