### 2021-11-11 8e5dca5
Add docs link to README and correct README instructions to build docs
