+++ docs/src/distributed.md
@@ -1,0 +1,187 @@
+# Running Distributed Simulations
+
+## Generating input files
+
+To use QXContexts we first generate a set of input files using the `generate_simulation_files`
+function.
+This function accepts any circuit type for which the `convert_to_tnc` function has been implemented.
+This includes QXZoo and YaoBlocks circuits, see the [Circuits with QXZoo and YaoQX](circuits.md)
+tutorial for more details on creating such circuits.
+The following example will use a 49 qubit Random Quantum Circuit (RQC) of depth 16.
+
+```
+using QXTools
+
+circ = create_rqc_circuit(7, 7, 16, 42, final_h=true)
+output_args = output_params_dict(49, 100)
+generate_simulation_files(circ, "rqc_7_7_16", time=10, output_args=output_args)
+```
+
+The above example will run the contraction finding algorithm for 10 seconds and write input files
+`rqc_7_7_16.qx`, `rqc_7_7_16.yaml` and `rqc_7_7_16.jld2`.
+By default two bonds will be sliced and the simulation files will specify that simulation should be run for 100 random bitstrings. Further details on the format and contents of these files is described in the [User Guide](users_guide.md) and further details of the slicing and output sampling options of the `generate_simulation_files` function are described in the [Miscellaneous Features](features.md) tutorial.
+
+## Running the simulation
+
+Once the input files have been generated, simulations can be run using the `bin/qxrun.jl` script which can be run directly from the command line (not the Julia REPL).
+To run a simulation using the input files generated in the example above one would use
+
+```
+julia --project bin/qxrun.jl -d rqc_7_7_16.qx -o rqc_7_7_16_output.jld2
+```
+
+This will run the simulation and write the output to the `rqc_7_7_16_output.jld2` file.
+
+To run simulations in parallel the MPI.jl package must first be installed and the `mpiexecjl` utility
+installed.
+This can be done with the following commands:
+
+```
+]add MPI
+import MPI
+MPI.install_mpiexecjl()
+```
+
+On compute clusters it is advised to use the system provided MPI installation. For more details on this see the official MPI.jl documentation [here](https://juliaparallel.github.io/MPI.jl/stable/configuration/#Using-a-system-provided-MPI).
+
+Once MPI.jl has been installed and configured the simulation example described above can be run in parallel on two processes with
+
+```
+mpiexecjl --project -n 2 julia -d rqc_7_7_16.qx -o rqc_7_7_16_output.jld2 -m
+```
+
+where the `-n 2` specifies the number of processes to use and the `-m` option enabled MPI.
+
+## Measuring the time and speedup
+
+When running the above example we would expect to see a reduction in the time taken when using more processes, however this is not what we observe in practice.
+This is due to the fact that with Julia when running code for the first time the majority of the time is taken up by compilation.
+In the above example since we are only calculating amplitudes for 10 bitstrings there are not enough computations after this compilation to observe a speedup.
+In practice for circuits of this size one would be calculating amplitudes for millions of bitstrings.
+It is also possible to reduce this startup time by compiling a custom system image as decribed in [this documentation](https://juliaqx.github.io/QXContexts.jl/dev/#Custom-system-image) for QXContexts.
+
+To see more clearly the time taken for compilation vs computation one can run the examples above using the `-t` flag. When this flag is used the simulation will be performed twice and timing information collected from each of these runs.
[…]
