+++ Project.toml
@@ -20,6 +20,7 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 TensorOperations = "6aa20fa7-93e2-5fca-9bc0-fbd0db3c71a2"
 TestSetExtensions = "98d24dd4-01ad-11ea-1b02-c9a08f80db04"
+TimerOutputs = "a759f4b9-e2f1-59dc-863e-4aeb61b1ea8f"
 YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
 
 [compat]
@@ -36,6 +37,7 @@ QXZoo = "0.1"
 Reexport = "1.1"
 TensorOperations = "3.1"
 TestSetExtensions = "2.0"
+TimerOutputs = "0.5"
 YAML = "0.4"
 julia = "1.6"
 
