### 2021-11-19 2775dcb
Add tutorials, minor changes to some default parameters and exported functions
