+++ docs/src/features.md
@@ -1,0 +1,127 @@
+# Miscellaneous Features
+
+In this tutorial we cover some of the features `QXTools` has to improve quantum circuit simulations and how to specify the type of output we would like our simulations to generate.
+
+## Hypergraphs
+
+One feature that `QXTools` uses by default is that of [hypergraphs](https://arxiv.org/abs/1710.05867). The hypergraph for a given tensor network is a graph, representing the network topology, whose nodes correspond to tensors in the network and edges correspond to indices in the network that connect different tensors. Note, two indices in the netwrok may be associated with the same edge in the hypergraph if they both connect to a diagonal tensor and are identical as a result. The tensor network topology can also be represented using an ordinary graph but it typically requires a greater number of edges to do so.
+
+This is useful because to find an efficient contraction plan for a given tensor network, `QXTools` first creates a line graph for the network topology and then runs the FlowCutter algorithm to find a tree decompostion for that line graph which can be converted into a contraction plan. Running FlowCutter on smaller graphs often results in more efficient contraction plans, with a smaller treewidth, in a shorter amount of time.
+
+`QXTools`, together with `QXTns`, can automatically detect the diagonal structure of gate tensors in a network and create the line graph of a network while taking the hypergraph structure of the network into account. We illustrate this below by creating the line graph for a network both with and without taking the hypergraph structure into account and show that using the hypergraph structure results in a smaller line graph.
+
+```
+# Create a random circuit.
+tnc = TensorNetworkCircuit(3)
+CX = [[1., 0., 0., 0.] [0., 1., 0., 0.] [0., 0., 0., 1.] [0., 0., 1., 0.]]
+
+for layer = 1:10
+    push!(tnc, [2, 1], CX)
+    push!(tnc, [2, 3], CX)
+    push!(tnc, [2], rand(2, 2))
+end
+
+# Create the line graph for the circuit without using the hypergraph.
+lg, symbol_map = convert_to_line_graph(tnc, use_hyperedges=false)
+@show lg
+
+# Create the line graph for the circuit using the hypergraph.
+lg, symbol_map = convert_to_line_graph(tnc, use_hyperedges=true)
+@show lg
+```
+
+
+## Slicing
+
+For simulations that use a contraction plan with a large treewidth, it may be the case that the simulation requires more memory than is available on the machine running the simulation. In this situation, it is necessary to reduce the memory requirements of the simulation.
+
+The standard method for reducing the memory requirements of a simulation in exchange for greater time complexity is known as slicing. This method works by fixing the value of selected indices in the network to decompose the contraction of the network into the sum of contractions of smaller networks. For more details on this method see the section on slicing in [this](https://arxiv.org/pdf/2005.06787.pdf) article.
+
+For a given contraction plan for a tensor network, `QXTools` uses [Shutski's tree-trimming](https://journals.aps.org/pra/abstract/10.1103/PhysRevA.102.062614), which is implemented in the package `QXGraphDecompositions` to select efficient indices of a network to slice. 
+
+To find a contraction plan for a quantum circuit using flowcutter and then slice a specific number of indices using Shutski's tree trimming method, the convenience function `contraction_scheme` can be used and is demonstrated below.
+
+```
+# Create a dummy circuit to test on.
+N = 5
+CX = [[1., 0., 0., 0.] [0., 1., 0., 0.] [0., 0., 0., 1.] [0., 0., 1., 0.]]
+tnc = TensorNetworkCircuit(N)
+for layer = 1:10, qubit = 1:N-1
+    push!(tnc, [qubit, qubit+1], CX)
+end
+add_input!(tnc)
+
[…]
