+++ paper.bib
@@ -43,3 +43,47 @@ We introduce \${\textbackslash}texttt\{Yao\}\$, an extensible, efficient open-so
 	pages = {341},
 	file = {Full Text PDF:/Users/nmoran/Zotero/storage/WMXHP8NQ/Luo et al. - 2020 - Yao.jl Extensible, Efficient Framework for Quantu.pdf:application/pdf;Snapshot:/Users/nmoran/Zotero/storage/4KGZ3URH/q-2020-10-11-341.html:text/html},
 }
+
+@article{Biamonte:2017,
+title={Tensor Networks in a Nutshell},
+author={Jacob Biamonte and Ville Bergholm},
+year={2017},
+eprint={1708.00006},
+archivePrefix={arXiv},
+primaryClass={quant-ph}
+}
+
+@article{Bridgeman:2017,
+title={Hand-waving and interpretive dance: an introductory course on tensor networks},
+volume={50},
+ISSN={1751-8121},
+url={http://dx.doi.org/10.1088/1751-8121/aa6dc3},
+DOI={10.1088/1751-8121/aa6dc3},
+number={22},
+journal={Journal of Physics A: Mathematical and Theoretical},
+publisher={IOP Publishing},
+author={Bridgeman, Jacob C and Chubb, Christopher T},
+year={2017},
+month={May},
+pages={223001}
+}
+
+@article{Wood:2011,
+title={Tensor networks and graphical calculus for open quantum systems},
+author={Christopher J. Wood and Jacob D. Biamonte and David G. Cory},
+year={2011},
+eprint={1111.6950},
+archivePrefix={arXiv},
+primaryClass={quant-ph}
+}
+
+@article{Villalonga:2019,
+author = {Villalonga, Benjamin and Lyakh, Dmitry and Boixo, Sergio and Neven, Hartmut and Humble, Travis S. and Biswas, Rupak and Rieffel, Eleanor G. and Ho, Alan and Mandr{\`{a}}, Salvatore},
+pages = {1--11},
+title = {{Establishing the Quantum Supremacy Frontier with a 281 Pflop/s Simulation}},
+url = {http://arxiv.org/abs/1905.00444},
+year = {2019},
+archivePrefix = "arXiv",
+eprint        = "1905.00444",
+primaryClass  = "quant-ph",
+}
