+++ paper.bib
@@ -87,3 +87,11 @@ archivePrefix = "arXiv",
 eprint        = "1905.00444",
 primaryClass  = "quant-ph",
 }
+
+@inproceedings{Hamann:2016,
+  author =       {Hamann, Michael and Strasser, Ben},
+  title =        {Graph Bisection with Pareto-Optimization},
+  publisher =    {ACM Journal of Experimental Algorithmics},
+  pages =        {1--34},
+  doi =          {10.1145/3173045},
+}
