### 2021-11-22 d01c1de
Add minor updates to examples
