+++ docs/src/features.md
@@ -8,7 +8,7 @@ One feature that `QXTools` uses by default is that of [hypergraphs](https://arxi
 
 This is useful because to find an efficient contraction plan for a given tensor network, `QXTools` first creates a line graph for the network topology and then runs the FlowCutter algorithm to find a tree decompostion for that line graph which can be converted into a contraction plan. Running FlowCutter on smaller graphs often results in more efficient contraction plans, with a smaller treewidth, in a shorter amount of time.
 
-`QXTools`, together with `QXTns`, can automatically detect the diagonal structure of gate tensors in a network and create the line graph of a network while taking the hypergraph structure of the network into account. We illustrate this below by creating the line graph for a network both with and without taking the hypergraph structure into account and show that using the hypergraph structure results in a smaller line graph.
+`QXTools`, together with `QXTns`, can automatically detect the diagonal structure of gate tensors in a network and create the line graph of a network while taking the hypergraph structure of the network into account. We illustrate this below by creating the line graph for a network both with and without taking the hypergraph structure into account and show that using the hypergraph structure results in a smaller line graph. Below, the line graph is represented by the [LabeledGraph](https://juliaqx.github.io/QXGraphDecompositions.jl/dev/LabeledGraph/) data structure implemented by the `QXGraphDecompositions` package. `@show`-ing this object should display the number of vertices and edges in the graph. 
 
 ```
 # Create a random circuit.
@@ -22,12 +22,13 @@ for layer = 1:10
 end
 
 # Create the line graph for the circuit without using the hypergraph.
+# The number of vertices and edges are displayed respectively.
 lg, symbol_map = convert_to_line_graph(tnc, use_hyperedges=false)
-@show lg
+@show lg.graph
 
 # Create the line graph for the circuit using the hypergraph.
 lg, symbol_map = convert_to_line_graph(tnc, use_hyperedges=true)
-@show lg
+@show lg.graph
 ```
 
 
@@ -62,11 +63,11 @@ metadata["Slicing"]["Treewidths after slicing consecutive edges"])
 
 ## QXContexts output
 
-The output of a distributed simulation run by `QXContexts` is determined by output parameters in the simulation files generated by the `generate_simulation_files` function in `QXTools`. To dictate which output is generated by `QXContexts`, a dictionary containing paramters for the desired output method needs to be passed to `generate_simulation_files`. To this end, `QXTools` provides the function `output_params_dict` which will create the relevant dictionary for the desired output. The tree output methods available in QXContexts are described below. Note, for each method, `output_params_dict` takes two positional arguments: `num_qubits` and `num_outputs` giving the number of qubits in the circuit to be simulated and the desired number of outputs to generate (ie number of bitstring samples or number of amplitudes to compute).
+The output of a distributed simulation run by `QXContexts` is determined by output parameters in the simulation files generated by the `generate_simulation_files` function in `QXTools`. To dictate which output is generated by `QXContexts`, a dictionary containing parameters for the desired output method needs to be passed to `generate_simulation_files`. To this end, `QXTools` provides the function `output_params_dict` which will create the relevant dictionary for the desired output. The tree output methods available in QXContexts are described below. Note, for each method, `output_params_dict` takes two positional arguments: `num_qubits` and `num_outputs` giving the number of qubits in the circuit to be simulated and the desired number of outputs to generate (ie number of bitstring samples or number of amplitudes to compute).
 
 ### Rejection
 
-To generate random bitstrings, which are distributed as if they were measured on a real quantum device running our quantum circuit, rejection sampling can be used. To create the output dictionary for rejection sampling, we can use the following keyword arguments of `output_params_dict`. 
[…]
