### 2021-12-08 daad680
33 add intro tutorials (#34)
* Add docs link to README and correct README instructions to build docs

* Add LiveServer to docs deps

* Add tutorials, minor changes to some default parameters and exported functions

* Add minor updates to examples

* Resolve comments

* Bump version number
