+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"users_guide/#User's-Guide","page":"User's Guide","title":"User's Guide","text":"","category":"section"},{"location":"users_guide/","page":"User's Guide","title":"User's Guide","text":"QXTools generates output files which provide a description of the computations and data required to perform a simulation. There are three output files:","category":"page"},{"location":"users_guide/","page":"User's Guide","title":"User's Guide","text":"Parameter file: This YAML file provides information on the sampling method to use for the simulation\nDSL file: The DSL file contains instructions describing the tensor operations involved in","category":"page"},{"location":"users_guide/","page":"User's Guide","title":"User's Guide","text":"performing the simulation. Uses .qx suffix","category":"page"},{"location":"users_guide/","page":"User's Guide","title":"User's Guide","text":"Data file: The data file contains the numerical values of the initial tensors of the tensor network","category":"page"},{"location":"users_guide/#Parallel-Processing","page":"User's Guide","title":"Parallel Processing","text":"","category":"section"},{"location":"users_guide/","page":"User's Guide","title":"User's Guide","text":"There are two levels of parallelism that are used to distribute computations over different processing elements. The highest level of parallism divides the set of output bitstrings among the available processing elements. The lower level of parallelism breaks each contraction into a sum of smaller contractions by slicing bonds of the tensor network. Slicing also reduces the total memory requirements and make it possible to perform contractions that would otherwise be intractable on a given system. However for each additional edge that is sliced, the network must be contracted for all possible values of the edge, leading to exponential growth in the number of contractions with the number of edges that are sliced. It is thus very important to choose the right edges to slice. This is a non trivial optimisation problem, but there exist many methods to which can find reasonable sets of edges to slice. In QXTools, the edges to slice are found when finding the contraciton plan which is handled by the QXGraphDecompositions package. Once the edges to select have been identified these appear as view commands in the DSL file, explained further below.","category":"page"},{"location":"users_guide/#Parameter-File","page":"User's Guide","title":"Parameter File","text":"","category":"section"},{"location":"users_guide/","page":"User's Guide","title":"User's Guide","text":"The parameter file parameters for the sampling method used to select output bitstrings. The following example shows how an explicit list of 3 bitstrings can be provided.","category":"page"},{"location":"users_guide/","page":"User's Guide","title":"User's Guide","text":"output:\n  method: List\n  params:\n    num_samples: 3\n    bitstrings:\n      - \"010\"\n      - \"101\"\n      - \"111\"","category":"page"},{"location":"users_guide/#DSL-Specification","page":"User's Guide","title":"DSL Specification","text":"","category":"section"},{"location":"users_guide/","page":"User's Guide","title":"User's Guide","text":"A Domain Specific Language (DSL) has been defined when enables better separation of concerns between components. This DSL describes a contraction over a tensor network to calculate observables of interest in the form of a sequence of instructions which act on individual tensors. This makes it possible to separate the development of the high performance distributed tensor network computation code from that of the higher level contraction planning, circuit and and network manipulation code. The \".qx\" suffix is used to identify these DSL files. We will first show a simple example of such a DSL file and then descibe each instruction in detail.","category":"
[…]
