### 2021-12-12 66e2fa5
Add tutorials, minor changes to some default parameters and exported functions, add qxrun.jl script (#37)
Co-authored-by: John Brennan <john.brennan@ichec.ie>
