### 2021-12-14 a27fbdf
Fix minor typos
