### 2021-12-15 a16bfda
Update to JOSS paper (#35)
* Add background section

* Add reference to flowcutter

Co-authored-by: John Brennan <john.brennan@ichec.ie>
