+++ paper.md
@@ -8,6 +8,7 @@ tags:
 authors:
 
   - name: John Brennan
+    orcid: 0000-0002-4428-6798
     affiliation: 1
   - name: Lee O’Riordan
     affiliation: 1
@@ -37,14 +38,23 @@ bibliography: paper.bib
 # Summary
 
 QXTools is a framework for simulating quantum circuits using tensor network methods.
-It is designed to run on large distributed compute clusters and supports GPU
-accelerators. The simulation workflow is broken down into a number of stages, each
-of which is managed by a special purpose package which can be used independently or
-as part of the QXTools framework. A domain specific language (DSL) is used to
-represent the simulation as a set of tensor network operations.
-This separates the high level index accounting and contraction planning from the
-low level implementation of the tensor network operations and makes it easier to
-support new hardware and network architectures.
+Weak simulation is the primary use case where given a quantum circuit and input state
+QXTools will efficiently calculate the probability amplitude of a given output configuration
+or set of configurations. Given this ability one can sample from the output distribution using
+random sampling approaches. QXTools is intended to be used by researchers interested in
+simulating circuits larger than those possible with full wave-function simulators or those
+interested in tensor network circuit simulation methods.
+
+QXTools is written in Julia [@Bezanson:2017] and is designed to run on large distributed 
+compute clusters and to support GPU accelerators. The simulation workflow is broken down into 
+a number of stages, each of which is managed by a special purpose package which can be used 
+independently or as part of the QXTools framework. To find efficient contraction orders for 
+tensor networks, an algorithm called FlowCutter [Hamann:2016] is used to construct tree 
+decompositions with optimal treewidth of the network's line graph by iteratively partitioning 
+it using maximal flows on the graph. A domain specific language (DSL) is used to express the 
+simulation as a set of tensor network operations. This separates the high level index 
+accounting and contraction planning from the low level implementation of the tensor network 
+operations and makes it easier to support new hardware and network architectures.
 
 # Statement of need
 
@@ -53,15 +63,21 @@ being run on them grow in complexity, simulations of these systems become much
 more computationally demanding. To reduce the turnaround time and
 allow larger systems to be simulated it is necessary to move beyond single workstations
 and use distributed compute clusters. QXTools provides a flexible, extensible open source
-framework for performing these simulations. The use of Julialang
+framework for performing these simulations. The use of Julia
 [@Bezanson:2017] makes it easy for the code
 to be understood, modified and extended while not sacrificing performance compared to
 compiled languages.
 
+# Background
+
[…]
