+++ paper.bib
@@ -94,4 +94,103 @@ primaryClass  = "quant-ph",
   publisher =    {ACM Journal of Experimental Algorithmics},
   pages =        {1--34},
   doi =          {10.1145/3173045},
+  year = {2018}
+}
+
+@article{Brennan:2021,
+	title = {Tensor {Network} {Circuit} {Simulation} at {Exascale}},
+	url = {http://arxiv.org/abs/2110.09894},
+	abstract = {Tensor network methods are incredibly effective for simulating quantum circuits. This is due to their ability to efficiently represent and manipulate the wave-functions of large interacting quantum systems. We describe the challenges faced when scaling tensor network simulation approaches to Exascale compute platforms and introduce QuantEx, a framework for tensor network circuit simulation at Exascale.},
+	urldate = {2021-12-14},
+	journal = {arXiv:2110.09894 [quant-ph]},
+	author = {Brennan, John and Allalen, Momme and Brayford, David and Hanley, Kenneth and Iapichino, Luigi and O'Riordan, Lee J. and Doyle, Myles and Moran, Niall},
+	month = oct,
+	year = {2021},
+	note = {arXiv: 2110.09894},
+	keywords = {Quantum Physics},
+	file = {arXiv Fulltext PDF:/Users/nmoran/Zotero/storage/PDGXHGPG/Brennan et al. - 2021 - Tensor Network Circuit Simulation at Exascale.pdf:application/pdf;arXiv.org Snapshot:/Users/nmoran/Zotero/storage/QDJ7EDKT/2110.html:text/html},
+}
+
+@article{Brayford:2021,
+	title = {Deploying {Containerized} {QuantEx} {Quantum} {Simulation} {Software} on {HPC} {Systems}},
+	url = {http://arxiv.org/abs/2110.05162},
+	abstract = {The simulation of quantum circuits using the tensor network method is very computationally demanding and requires significant High Performance Computing (HPC) resources to find an efficient contraction order and to perform the contraction of the large tensor networks. In addition, the researchers want a workflow that is easy to customize, reproduce and migrate to different HPC systems. In this paper, we discuss the issues associated with the deployment of the QuantEX quantum computing simulation software within containers on different HPC systems. Also, we compare the performance of the containerized software with the software running on bare metal.},
+	urldate = {2021-12-14},
+	journal = {arXiv:2110.05162 [cs]},
+	author = {Brayford, David and Brennan, John and Allalen, Momme and Hanley, Kenneth and Iapichino, Luigi and ORiordan, Lee and Moran, Niall},
+	month = oct,
+	year = {2021},
+	note = {arXiv: 2110.05162},
+	keywords = {Computer Science - Distributed, Parallel, and Cluster Computing},
+	file = {arXiv Fulltext PDF:/Users/nmoran/Zotero/storage/3EWDFN2G/Brayford et al. - 2021 - Deploying Containerized QuantEx Quantum Simulation.pdf:application/pdf;arXiv.org Snapshot:/Users/nmoran/Zotero/storage/5MVWSDFR/2110.html:text/html},
+}
+
+@misc{Torlai:2020,
+    title={\mbox{PastaQ}: A Package for Simulation, Tomography and Analysis of Quantum Computers},
+    author={Giacomo Torlai and Matthew Fishman},
+    year={2020},
+    url={https://github.com/GTorlai/PastaQ.jl/}
+}
+
+
+@article{Gray:2018,
+	title = {quimb: {A} python package for quantum information and many-body calculations},
+	volume = {3},
+	issn = {2475-9066},
+	shorttitle = {quimb},
+	url = {https://joss.theoj.org/papers/10.21105/joss.00819},
+	doi = {10.21105/joss.00819},
[…]
