+++ paper.md
@@ -43,17 +43,19 @@ QXTools will efficiently calculate the probability amplitude of a given output c
 or set of configurations. Given this ability one can sample from the output distribution using
 random sampling approaches. QXTools is intended to be used by researchers interested in
 simulating circuits larger than those possible with full wave-function simulators or those
-interested in tensor network circuit simulation methods.
+interested in research and development of tensor network circuit simulation methods.
+See [@Brennan:2021] for more complete background and scaling results and [@Brayford:2021] for
+details about deploying in containerised environments.
 
-QXTools is written in Julia [@Bezanson:2017] and is designed to run on large distributed 
-compute clusters and to support GPU accelerators. The simulation workflow is broken down into 
-a number of stages, each of which is managed by a special purpose package which can be used 
-independently or as part of the QXTools framework. To find efficient contraction orders for 
-tensor networks, an algorithm called FlowCutter [Hamann:2016] is used to construct tree 
-decompositions with optimal treewidth of the network's line graph by iteratively partitioning 
-it using maximal flows on the graph. A domain specific language (DSL) is used to express the 
-simulation as a set of tensor network operations. This separates the high level index 
-accounting and contraction planning from the low level implementation of the tensor network 
+QXTools is written in Julia [@Bezanson:2017] and is designed to run on large distributed
+compute clusters and to support GPU accelerators. The simulation workflow is broken down into
+a number of stages, each of which is managed by a special purpose package which can be used
+independently or as part of the QXTools framework. To find efficient contraction orders for
+tensor networks, an algorithm called FlowCutter [@Hamann:2016] is used to construct tree
+decompositions with optimal treewidth of the network's line graph by iteratively partitioning
+it using maximal flows on the graph. A domain specific language (DSL) is used to express the
+simulation as a set of tensor network operations. This separates the high level index
+accounting and contraction planning from the low level implementation of the tensor network
 operations and makes it easier to support new hardware and network architectures.
 
 # Statement of need
@@ -70,9 +72,13 @@ compiled languages.
 
 # Background
 
-Classical simulation of quantum circuits is essential for debugging and validating the accuracy of quantum computing devices and algorithms. This is a very computationally demanding problem owing to the exponential growth in the state space as the number of qubits is increased. Up until recently it has been possible to simulate the largest prototype universal quantum computers using direct evolution of the quantum state (where the full wave-function is stored in memory (or on disk)) using modest personal computing resources. With the recent emergence of Noisy Intermediate Scale Quantum (NISQ) devices, it has become intractable to use this approach to simulate devices of this size on even the largest supercomputers. These advances have necessitated the development of new circuit simulation methods which can simulate large systems without incurring the memory required to store the full wave-function. Tensor network methods have been demonstrated to achieve state of the art performance in simulating Random Quantum Circuits (RQC) as part of the quantum sumpremacy experiments [@Villalonga:2019]. Despite the impressive results achieved with these methods to date they are not suitable for all types of circuits and in many cases full wave-function methods are preferable. For example for highly entangled circuits, the tensor network representation will consume the same memory as full wave-function methods but will incur additional overhead. However for circuits with moderate entanglement and cases where one is not interested in exact results, but in results up to a particulary fidelity, tensor network approaches can offer significant advantages.
[…]
