### 2021-12-16 #3711 @nmoran
Thanks again to @goerz and @obliviateandsurrender for their time and efforts in reviewing our work. Over the past few weeks we have made significant efforts to address the reviewers comments with the following changes:

- Addition of four tutorials which run through examples and explanations of different components and expand on the simple example included previously
- Correction of minor issues highlighted including adding link to documentation, fixing link to contributing guide and adding an issue template
- Addition of statement of need and background sections to the paper text which include reference to other tools and mention of distinguishing features. References to other recent papers about the toolkit are also included which feature more background and some performance results
