### 2022-01-06 1182681
Update deps, version number and some tests
