+++ Project.toml
@@ -1,7 +1,7 @@
 name = "QXTools"
 uuid = "84f0eee1-10ae-40da-994b-b9d0e53829ae"
 authors = ["QuantEx team"]
-version = "0.2.5"
+version = "0.2.3"
 
 [deps]
 AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
@@ -30,7 +30,7 @@ DataStructures = "0.18"
 FileIO = "1.6"
 JLD2 = "0.4"
 LightGraphs = "1.3"
-QXContexts = "0.2"
+QXContexts = "0.3"
 QXGraphDecompositions = "0.2"
 QXTns = "0.1"
 QXZoo = "0.1"
