### 2022-01-06 17d2966
Update deps, version number and some tests (#42)
