### 2022-01-06 JuliaQX/QXTools.jl#43 @goerz
Statement of Need
From https://github.com/openjournals/joss-reviews/issues/3711

Even though I'm not sure I agree, JOSS is very explicit in their requirements that there has to be a "statement of need" *in the online documentation*. So, can you incorporate the "Statement of need" from the JOSS manuscript somewhere in the documentation, so that I can check off this item?

Incidentally, the "Docs/stable" button in the main README is broken (404 response). I would actually recommend setting up [Documentation Versions](https://juliadocs.github.io/Documenter.jl/stable/man/hosting/#Documentation-Versions) for your releases, which would fix this.
