### 2022-01-06 JuliaQX/QXTools.jl#43 @goerz
The "Docs/stable" buttons are also broken in the various sub-projects like QXTns.jl
