### 2022-01-06 JuliaQX/QXTools.jl#44 @goerz
Link to documentation of sub-projects
For https://github.com/openjournals/joss-reviews/issues/3711

E.g. on https://juliaqx.github.io/QXTools.jl/dev/#QXTools, wherever the various super-/sub-projects are mentioned (QuantEx, QXZoo, QXTns, etc.), can you make these links to the docs of those packages? It would make it much easier to navigate the full project documentation, e.g. to get to the API docs

It also wouldn't hurt to turn any mention of external packages (" ITensors, TensorOperations.jl, OMEinsum.jl, Yao.jl, TAL-SH and ExaTN") into links to those projects.
