### 2022-01-06 #3711 @goerz
Ok, I finished my second round of review. The manuscript and documentation seems good enough for publication, as far as I'm concerned. I opened two relatively trivial issues, https://github.com/JuliaQX/QXTools.jl/issues/43 (which is a pure formality) and https://github.com/JuliaQX/QXTools.jl/issues/44.

I would recommend making an official release of the package (maybe even a 1.0 release - In my mind, publishing a paper for a software package should be bit of a "stable" stamp) immediately before the final publication so that the paper can be tied to that version. I would also recommend setting up [versioned docs](https://juliadocs.github.io/Documenter.jl/stable/man/hosting/#Documentation-Versions). This would apply to all of the packages tied together with QXTools.
