### 2022-01-09 2d0be62
Separate CI workflow for docs
