### 2022-01-09 e339195
Separate CI workflow for docs (#45)
