### 2022-01-09 JuliaQX/QXTools.jl#43 @nmoran
Update so documentation workflow runs for tags in #45
