### 2022-01-09 JuliaQX/QXTools.jl#43 @JuliaRegistrator
Registration pull request created: [JuliaRegistries/General/51980](https://github.com/JuliaRegistries/General/pull/51980)

After the above pull request is merged, it is recommended that a tag is created on this repository for the registered package version.

This will be done automatically if the [Julia TagBot GitHub Action](https://github.com/marketplace/actions/julia-tagbot) is installed, or can be done manually through the github interface, or via:
```
git tag -a v0.2.4 -m "<description of version>" e3391958605a380e1e12232ddb16538a3f6685bc
git push origin v0.2.4
```
