### 2022-01-10 c84a66c
Add statement of need to docs and links to subpackages
