### 2022-01-10 6c5f2a8
Add statement of need to docs and links to subpackages (#47)
* Add statement of need to docs and links to subpackages

* Bump patch number
