### 2022-01-10 JuliaQX/QXTools.jl#44 @JuliaRegistrator
Registration pull request created: [JuliaRegistries/General/52061](https://github.com/JuliaRegistries/General/pull/52061)

After the above pull request is merged, it is recommended that a tag is created on this repository for the registered package version.

This will be done automatically if the [Julia TagBot GitHub Action](https://github.com/marketplace/actions/julia-tagbot) is installed, or can be done manually through the github interface, or via:
```
git tag -a v0.2.5 -m "<description of version>" 6c5f2a89e2a15f976cfb8faa49a8f19d3504bc84
git push origin v0.2.5
```
