### 2022-01-18 #3711 @nmoran
@goerz thanks for reviewing our updates and the further suggestions. We have addressed the issues you raised with the documentation links.

@jarvist while we hope to continue making updates and improvements over time, there are no outstanding edits or changes at this time.
