### 2022-01-18 #3711 @obliviateandsurrender
Thank you @nmoran for the timely improvements. Things look great from my end as well but just one minor comment:
1. Citation for PastaQ library in the [manuscript](https://github.com/openjournals/joss-papers/blob/joss.03711/joss.03711/10.21105.joss.03711.pdf) appears broken to me. You may want to use the following from their GitHub repo:
```
@misc{pastaq,
    title={\mbox{PastaQ}: A Package for Simulation, Tomography and Analysis of Quantum Computers},
    author={Giacomo Torlai and Matthew Fishman},
    year={2020},
    url={https://github.com/GTorlai/PastaQ.jl/}
}
```
Other than this, I believe the package is now ready for publication and also for a stable release!
