### 2022-01-20 0542ea5
Update ci to use julia 1.7 (#48)
