+++ Project.toml
@@ -39,7 +39,7 @@ TensorOperations = "3.1"
 TestSetExtensions = "2.0"
 TimerOutputs = "0.5"
 YAML = "0.4"
-julia = "1.6"
+julia = "1.7"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
