+++ paper.bib
@@ -125,7 +125,7 @@ primaryClass  = "quant-ph",
 	file = {arXiv Fulltext PDF:/Users/nmoran/Zotero/storage/3EWDFN2G/Brayford et al. - 2021 - Deploying Containerized QuantEx Quantum Simulation.pdf:application/pdf;arXiv.org Snapshot:/Users/nmoran/Zotero/storage/5MVWSDFR/2110.html:text/html},
 }
 
-@misc{Torlai:2020,
+@misc{pastaq,
     title={\mbox{PastaQ}: A Package for Simulation, Tomography and Analysis of Quantum Computers},
     author={Giacomo Torlai and Matthew Fishman},
     year={2020},
