+++ Project.toml
@@ -33,7 +33,7 @@ LightGraphs = "1.3"
 QXContexts = "0.3"
 QXGraphDecompositions = "0.2"
 QXTns = "0.1"
-QXZoo = "0.1"
+QXZoo = "0.1, 1"
 Reexport = "1.1"
 TensorOperations = "3.1"
 TestSetExtensions = "2.0"
