### 2022-01-23 4c9d626
CompatHelper: bump compat for QXGraphDecompositions to 1, (keep existing compat)
