+++ Project.toml
@@ -31,7 +31,7 @@ FileIO = "1.6"
 JLD2 = "0.4"
 LightGraphs = "1.3"
 QXContexts = "0.3"
-QXGraphDecompositions = "0.2"
+QXGraphDecompositions = "0.2, 1"
 QXTns = "0.1"
 QXZoo = "0.1"
 Reexport = "1.1"
