+++ Project.toml
@@ -30,7 +30,7 @@ DataStructures = "0.18"
 FileIO = "1.6"
 JLD2 = "0.4"
 LightGraphs = "1.3"
-QXContexts = "0.3"
+QXContexts = "0.3, 1"
 QXGraphDecompositions = "0.2"
 QXTns = "0.1"
 QXZoo = "0.1"
