+++ Project.toml
@@ -32,7 +32,7 @@ JLD2 = "0.4"
 LightGraphs = "1.3"
 QXContexts = "0.3"
 QXGraphDecompositions = "0.2"
-QXTns = "0.1"
+QXTns = "0.1, 1"
 QXZoo = "0.1"
 Reexport = "1.1"
 TensorOperations = "3.1"
