### 2022-01-26 163c3c7
Bump to 1.0.0 (#53)
