### 2022-01-26 JuliaQX/QXTools.jl#44 @JuliaRegistrator
Registration pull request created: [JuliaRegistries/General/53209](https://github.com/JuliaRegistries/General/pull/53209)

After the above pull request is merged, it is recommended that a tag is created on this repository for the registered package version.

This will be done automatically if the [Julia TagBot GitHub Action](https://github.com/marketplace/actions/julia-tagbot) is installed, or can be done manually through the github interface, or via:
```
git tag -a v1.0.0 -m "<description of version>" 163c3c742d7292eb5c192967610541db976de453
git push origin v1.0.0
```
