### 2022-02-15 #3711 @jarvist
@nmoran have you archived (to generate a DOI) your version v1.0.0 of the software?
This can be done via Zenodo, with the Github guide quite easy to follow https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content
