### 2022-02-15 #3711 @jarvist
OK, I've done my last proof reading and checks. Once we have a DOI for the software archive, I'm more than happy to recommend acceptance & pass it to the Editor in Chief for final checks and processing.
