### 2022-02-15 #3711 @jarvist
Hmm, I'm not sure if the DOI should be wrapped as a URL (with https://doi.org/) or left raw.
