### 2022-02-15 #3711 @danielskatz
FYI, the command to rerun the check is `@whedon check references from branch joss-paper`, and this can be done by the author
