+++ paper.bib
@@ -72,6 +72,7 @@ pages={223001}
 title={Tensor networks and graphical calculus for open quantum systems},
 author={Christopher J. Wood and Jacob D. Biamonte and David G. Cory},
 year={2011},
+DOI={10.26421/qic15.9-10-3},
 eprint={1111.6950},
 archivePrefix={arXiv},
 primaryClass={quant-ph}
@@ -83,6 +84,7 @@ pages = {1--11},
 title = {{Establishing the Quantum Supremacy Frontier with a 281 Pflop/s Simulation}},
 url = {http://arxiv.org/abs/1905.00444},
 year = {2019},
+doi = {10.1088/2058-9565/ab7eeb},
 archivePrefix = "arXiv",
 eprint        = "1905.00444",
 primaryClass  = "quant-ph",
@@ -105,6 +107,7 @@ primaryClass  = "quant-ph",
 	journal = {arXiv:2110.09894 [quant-ph]},
 	author = {Brennan, John and Allalen, Momme and Brayford, David and Hanley, Kenneth and Iapichino, Luigi and O'Riordan, Lee J. and Doyle, Myles and Moran, Niall},
 	month = oct,
+	doi = {10.1109/qcs54837.2021.00006},
 	year = {2021},
 	note = {arXiv: 2110.09894},
 	keywords = {Quantum Physics},
@@ -119,6 +122,7 @@ primaryClass  = "quant-ph",
 	journal = {arXiv:2110.05162 [cs]},
 	author = {Brayford, David and Brennan, John and Allalen, Momme and Hanley, Kenneth and Iapichino, Luigi and ORiordan, Lee and Moran, Niall},
 	month = oct,
+	doi = {10.1109/canopiehpc54579.2021.00005},
 	year = {2021},
 	note = {arXiv: 2110.05162},
 	keywords = {Computer Science - Distributed, Parallel, and Cluster Computing},
@@ -190,6 +194,7 @@ ADS Bibcode: 2019APS..MARB22012M},
 	author = {Pang, Yuchen and Hao, Tianyi and Dugad, Annika and Zhou, Yiqing and Solomonik, Edgar},
 	month = sep,
 	year = {2020},
+	doi = {10.1109/sc41405.2020.00018},
 	note = {arXiv: 2006.15234},
 	keywords = {Computer Science - Distributed, Parallel, and Cluster Computing, Physics - Computational Physics, Quantum Physics},
 	file = {arXiv Fulltext PDF:/Users/nmoran/Zotero/storage/VIHLM8IA/Pang et al. - 2020 - Efficient 2D Tensor Network Simulation of Quantum .pdf:application/pdf;arXiv.org Snapshot:/Users/nmoran/Zotero/storage/KFYANKD7/2006.html:text/html},
