+++ paper.bib
@@ -25,7 +25,7 @@
 }
 
 @article{Luo:2020,
-	title = {Yao.jl: {Extensible}, {Efficient} {Framework} for {Quantum} {Algorithm} {Design}},
+	title = {Yao.jl: Extensible, Efficient Framework for Quantum Algorithm Design},
 	volume = {4},
 	shorttitle = {Yao.jl},
 	url = {https://quantum-journal.org/papers/q-2020-10-11-341/},
@@ -71,8 +71,12 @@ pages={223001}
 @article{Wood:2011,
 title={Tensor networks and graphical calculus for open quantum systems},
 author={Christopher J. Wood and Jacob D. Biamonte and David G. Cory},
-year={2011},
+year={2015},
 DOI={10.26421/qic15.9-10-3},
+journal={Quantum Information and Computation},
+volume=15,
+number={9--10},
+month=jul,
 eprint={1111.6950},
 archivePrefix={arXiv},
 primaryClass={quant-ph}
@@ -81,7 +85,7 @@ primaryClass={quant-ph}
 @article{Villalonga:2019,
 author = {Villalonga, Benjamin and Lyakh, Dmitry and Boixo, Sergio and Neven, Hartmut and Humble, Travis S. and Biswas, Rupak and Rieffel, Eleanor G. and Ho, Alan and Mandr{\`{a}}, Salvatore},
 pages = {1--11},
-title = {{Establishing the Quantum Supremacy Frontier with a 281 Pflop/s Simulation}},
+title = {Establishing the Quantum Supremacy Frontier with a 281 Pflop/s Simulation},
 url = {http://arxiv.org/abs/1905.00444},
 year = {2019},
 doi = {10.1088/2058-9565/ab7eeb},
@@ -90,17 +94,18 @@ eprint        = "1905.00444",
 primaryClass  = "quant-ph",
 }
 
-@inproceedings{Hamann:2016,
+@article{Hamann:2016,
   author =       {Hamann, Michael and Strasser, Ben},
   title =        {Graph Bisection with Pareto-Optimization},
-  publisher =    {ACM Journal of Experimental Algorithmics},
+  journal =    {ACM Journal of Experimental Algorithmics},
   pages =        {1--34},
+  volume = 23,
   doi =          {10.1145/3173045},
   year = {2018}
 }
 
 @article{Brennan:2021,
-	title = {Tensor {Network} {Circuit} {Simulation} at {Exascale}},
+	title = {Tensor Network Circuit Simulation at Exascale},
 	url = {http://arxiv.org/abs/2110.09894},
 	abstract = {Tensor network methods are incredibly effective for simulating quantum circuits. This is due to their ability to efficiently represent and manipulate the wave-functions of large interacting quantum systems. We describe the challenges faced when scaling tensor network simulation approaches to Exascale compute platforms and introduce QuantEx, a framework for tensor network circuit simulation at Exascale.},
 	urldate = {2021-12-14},
@@ -115,7 +120,7 @@ primaryClass  = "quant-ph",
 }
 
 @article{Brayford:2021,
-	title = {Deploying {Containerized} {QuantEx} {Quantum} {Simulation} {Software} on {HPC} {Systems}},
+	title = {Deploying Containerized {QuantEx} Quantum Simulation Software on {HPC} Systems},
 	url = {http://arxiv.org/abs/2110.05162},
[…]
