### 2022-02-15 #3711 @danielskatz
I've also suggested a bunch of changes in the paper and bib in https://github.com/JuliaQX/QXTools.jl/pull/54
